Attach Manila shares via virtiofs (db)
This patch introduces the new table and associated migrations to support
share mapping between shares and instances.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: Id2956188eef2160e718f08d63ced81e9a25e47af