Attach Manila shares via virtiofs (manila abstraction)
This patch is inspired by /nova/volume/cinder.py, it is an abstraction to
the manila service.
Note: This path is currently using the python-manila client because some
methods (allow/deny shares) are not available in the openstacksdk.
It is planned to move to the sdk as soons as it will be ready.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Depends-On: https://review.opendev.org/c/openstack/openstacksdk/+/889519

Change-Id: I44ab37ec2c15fcfc351c42216660bda39461b163