Attach Manila shares via virtiofs (objects)
This patch introduce the ShareMapping and ShareMappingList top level objects
and associated attach and detach methods that interact with the DB
It also introduce ShareMappingLibvirt and ShareMappingLibvirtNFS
children objects that will be used by the driver part.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: Ibacfeaf7daa726bf0ee7802bacd3f70329fec624