Add metadata for shares
This patch add device metatada to an instance if a share is is attached.
Users can get the share information (share_id and tag) from the metadata
and create automation to mount the share on the VM.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: I4c7bb15dbe423ec2815efe9d9505b1e8f2a3c4f5