Check shares support (API)
This patch allows to block requests involving shares, if the compute API is not
at the correct version.
It also blocks instance actions not allowed (shelve, evacuate,...) if
the instance is running and attached to a share.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: Ic470178e91d6aa247892939e5d2475530b4f72b4