Add shares to InstancePayload
This patch add share information to the InstancePayload that can be sent
if the include_share_mapping configuration parameter is enabled.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: I3d5005eab9e3f23be149e955e8cb4608a6ee1312