Add libvirt test to ensure metadata are working.
This patch add libvirt functional test to ensure metadata are transmitted
up to the metadata service.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: I67c33f249ceb6cd51b6843acd347205777e89be0