Add instance.power_off_error notification
This patch add a notification when a share can not be umounted on the
compute host.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: If53862447d9afddab5cd4a33fb2964866129e489