Add share_info parameter to reboot method for each driver (driver part)
This patch allow to pass share_info parameter to the reboot
method. So it will allow to reboot instance and keep the
shares attached.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: I897cbd9ebc91938d0227d988b4bc833f2216618a