Add instance.share_detach_error notification
This patch add a notification when a share can not be unmounted on the
compute host.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: Ic1aeb2308af5bf645fa67fe1f47e5a076efd6071