Add share_info parameter to resume method for each driver (driver part)
This patch allow to pass share_info parameter to the resume
method. So it will allow to resume an instance and keep the
shares attached.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: I66da2b010caa7e7bed15e69c1e14fd22ca1c8a31