Mounting the shares as part of the initialization process
The purpose of this patch is to ensure that, in the event of a compute
reboot, shares associated with instances are mounted successfully on the
compute host during the service initialization process.

To use Manila as an admin, it is necessary to have a Manila section in the
"nova.conf" configuration file. As a result, the dependencies below have
been introduced in DevStack to satisfy this requirement.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Depends-On: Ia17c7a136cbe83efa1ef4e302d1c404034a50cda
Change-Id: Ia498c72e1d79e98e4d50e2337a5acc4921910f06