Manage local share (driver part)
This patch serves the purpose of managing local shares. It involves
modifying the libvirt XML file to establish a connection between the
VM and the local share. The tag used for this connection is determined
by the configuration value of the share_local_parameter parameter.

The share_local_parameter key specifies the root of the file system
(fs) and appends the instance name to define the source directory
that will be mapped.

Furthermore, the patch introduces support for the "readonly" attribute,
but it is currently inactive due to the lack of support for it in
virtiofs.

Implements: blueprint virtiofs-scaphandre
Change-Id: I6f4fb2a355b1078de519d92c53e30034aa0f1316