Manage local share (compute manager part)
The purpose of this patch is to enhance the functionality of
the _get_share_info method. It has been modified to include the
configuration of a local share, if necessary. Depending on the status,
the method now manages the share_mappings to determine whether a
local share should be added to the instance or not.

Upon execution, this method will provide the ShareMappingList object,
which defines the attachments to be associated with the instance.

Implements: blueprint virtiofs-scaphandre
Change-Id: Ib471fd31e41df226ad0f09c4328b33eb7d0506eb