Amend ShareMappingStatus due to asynchronous call
This update introduces the "attaching" status, which indicates
the beginning of a connection between a share and a server. When
this occurs, the compute system will permit a policy to connect
the share through Manila and will record this connection in the
share_mapping table. Once completed, the status will transition to
"inactive." Additionally, the "detaching" status has been added to
indicate the end of a link between a share and a server. During this
process, the compute system will revoke the share's policy and delete
its entry from the share_mapping table.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: I1724940474b2e96e0c3962095f77df55c749529a