Amend ShareMappingStatus due to asynchronous call
This update adds the "attaching" status, signifying the initiation
of a connection between a share and a server, and includes a period
of waiting until the share is mounted on the compute node. Once the
share is successfully mounted and operational, its status will change
to "inactive."

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: I1724940474b2e96e0c3962095f77df55c749529a