Amend DB model add a unique constraint.
This update adds a constraint between instance_uuid and share_id to
make sure it will be unique. It aims to avoid potential conflicts
from simultaneous API creation requests, which could result in the
creation of duplicate records in the database.

Manila is the OpenStack Shared Filesystems service. These series
of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances
using virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: I419fd40baf9acf487154eebc77c181633ea53272