Reports instance events to the DB regarding attaching and detaching a share
This patch fix the attaching and detaching events that were not saved
into the DB.

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow the shares
provided by Manila to be associated with and attached to instances using
virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares

Change-Id: Iee865517a08b84c70ba98617923367897f070eb9